multi: add network capacity based generator
When we want to simulate random activity in large network, keeping track
of our graph will be the most expensive part of the simulation. This is
because we'll want to cache node capacities for easy lookup, and
because a weighted index has a large memory footprint (N -1 u64s for
our use). To minimize the cost of tracking these values, we use a single
NetworkGraphView tracker that will provide expensive operations in a
single location.