multi: make expected payment amount and multiplier configurable
Surface so that end users can configure these values. Note that
expected payment amount is intentionally called "avg" in the user
config because that's a more user friendly / recognizable term, but
we keep expected payment amount elsewhere in the code base since it's
the expectation we provide for our distribution.