sim-lib: Redefine RandomActivityErrors to provide finer granularity
This also makes the errors for the `RandomActivityErrors` local, and `SimulationError`
use it. Same goes for the errors returned by the `PaymentGenerator`. They used to
use `RandomActivityError`, but if that's the case it would have made more sense to be
part of `random_activity.rs`. If we define it as something more general (that can therefore
be used for other things apart from random activity generation) it make more sense to use
a more general error.